refactor(core): isolate registry encoding utilities
